rpi_ds18b20_temperature_logger/pyproject.toml
[project]
name = "rpi-ds18b20-temperature-logger"
version = "0.1.0"
description = "RPI Temperature logger with DS18B20"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pydantic>=2.11.7",
    "pyyaml>=6.0.2",
]

[dependency-groups]
dev = []
